scrapers/ScrapingGenericScraper/requirements.txt
async-generator==1.10
attrs==22.2.0
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==40.0.1
cssselect==1.2.0
exceptiongroup==1.1.1
filelock==3.10.7
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==1.0.1
lxml==4.9.2
outcome==1.2.0
packaging==23.0
parsel==1.7.0
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==23.1.1
PySocks==1.7.1
python-dotenv==1.0.0
queuelib==1.6.2
requests==2.28.2
requests-file==1.5.1
Scrapy==2.8.0
selenium==4.8.3
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tldextract==3.4.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
Twisted==22.10.0
typing_extensions==4.5.0
urllib3==1.26.15
w3lib==2.1.1
webdriver-manager==3.8.5
wsproto==1.2.0
zope.interface==6.0
